fix(Dialog): 移除 maskClosable 条件下不必要的 onClose 存在性检查，确保拖拽行为不会意外触发对话框关闭。同时将相关测试从独立文件合并到主测试文件中以简化结构。
