+++ tests/index.spec.tsx
@@ -188,6 +188,30 @@ describe('dialog', () => {
     expect(onClose).not.toHaveBeenCalled();
   });
 
+  it('should not close when dragging from content to mask', () => {
+    const onClose = jest.fn();
+    const { getByText } = render(
+      <Dialog visible maskClosable onClose={onClose}>
+        Content
+      </Dialog>
+    );
+
+    jest.runAllTimers();
+
+    const content = getByText('Content');
+    const mask = document.querySelector('.rc-dialog-wrap');
+    if (!mask) throw new Error('Mask not found');
+
+    // Simulate mouse down on content
+    fireEvent.mouseDown(content);
+    // Simulate mouse up on mask
+    fireEvent.mouseUp(mask);
+    // Simulate click on mask (since click follows mouseup)
+    fireEvent.click(mask);
+
+    expect(onClose).not.toHaveBeenCalled();
+  });
+
   it('renderToBody', () => {
     const container = document.createElement('div');
     document.body.appendChild(container);
